backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
websockets
numpy
torch
torchaudio
pyannote.audio
pydub
python-dotenv
httpx
scikit-learn
sounddevice
